twister: share the status property via a StatusMixin
TestCase, TestSuite and TestInstance each defined a byte-identical
status property/setter validating assignments against TwisterStatus and
raising StatusAttributeError. Extract the shared logic into a StatusMixin
in statuses.py and have the three classes inherit it, dropping the now
unused Enum and StatusAttributeError imports.
