+++ scripts/pylib/twister/twisterlib/statuses.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 from enum import Enum
 
 from colorama import Fore
+from twisterlib.error import StatusAttributeError
 
 
 class TwisterStatus(str, Enum):
@@ -52,3 +53,26 @@ class TwisterStatus(str, Enum):
     FAIL = 'failed'
     PASS = 'passed'
     SKIP = 'skipped'
+
+
+class StatusMixin:
+    """Mixin providing a validated ``status`` property backed by ``_status``.
+
+    Assigning a value that is not a valid TwisterStatus (by name or value)
+    raises StatusAttributeError, so illegal statuses fail loudly at the
+    assignment site. Shared by TestCase, TestSuite and TestInstance.
+    """
+    _status: TwisterStatus = TwisterStatus.NONE
+
+    @property
+    def status(self) -> TwisterStatus:
+        return self._status
+
+    @status.setter
+    def status(self, value: TwisterStatus) -> None:
+        # Check for illegal assignments by value
+        try:
+            key = value.name if isinstance(value, Enum) else value
+            self._status = TwisterStatus[key]
+        except KeyError as err:
+            raise StatusAttributeError(self.__class__, value) from err
